Stereo separation now applied at replayers + search tab shows navbar until search gets focus

+++ 4champ/Replay/OpenMPT/MPTReplayer.m
@@ -32,7 +32,7 @@
 }
 
 - (void) setStereoSeparation:(NSInteger)value {
-  int newValue = value * 2; // (parameter range 0-100, mpt range 0-200)
+  int newValue = (int)value * 2; // (parameter range 0-100, mpt range 0-200)
   openmpt_module_set_render_param(currentOMPTFile, OPENMPT_MODULE_RENDER_STEREOSEPARATION_PERCENT, (int32_t)newValue);
 }
